+++ frontend/src/pages/QueryDocuments/stores/useDocTasksCurrent.ts
@@ -31,6 +31,8 @@ interface DocTasksCurrentState {
   setFromServerResponse: (response: DocTasksResponse) => void;
   needReload: boolean;
   setNeedReload: (needReload: boolean) => void;
+  opUUID: string;
+  setOpUUID: (opUUID: string) => void;
   setFromHistory: (item: DocTasksShortItem) => void;
   setNewQuery: () => void;
 }
@@ -54,6 +56,7 @@ const defaultState = {
   is_error: null,
   status_pct: null,
   needReload: false,
+  opUUID: "",
 };
 
 export const useDocTasksCurrentStore =
@@ -68,6 +71,7 @@ export const useDocTasksCurrentStore =
     setBeforeServerResponse: (query: DocTasksQuery) => set(query),
     setFromServerResponse: (response: DocTasksResponse) => set(response),
     setNeedReload: (needReload: boolean) => set({ needReload }),
+    setOpUUID: (opUUID: string) => set({ opUUID }),
     setFromHistory: (item: DocTasksShortItem) =>
       set({
         doc_task_id: item.doc_task_id,